Consolidate chart viz e2e specs (#43619)
